Move tests out of src
