+++ package.json
@@ -7,15 +7,15 @@
     "clean": "rm -Rf ./lib",
     "compile": "npm run clean && babel src --out-dir lib",
     "watch": "babel src --watch --out-dir lib",
-    "test": "jest --coverage ./lib",
+    "test": "jest --coverage",
     "start": "npm run lint && npm run compile && npm test",
-    "onchange": "onchange 'lib/*.js' 'lib/**/*.js' -- npm run lintAndTest",
+    "onchange": "onchange 'src/*.js' 'src/**/*.js' -- npm run lintAndTest",
     "dev": "npm run watch & npm run onchange",
     "prepublish": "npm ls && npm run lint && npm test",
     "example:compile": "gulp",
     "example:watch": "gulp watch:example",
     "lint": "eslint ./src",
-    "lintAndTest": "npm run lint && jest ./lib"
+    "lintAndTest": "npm run lint && jest"
   },
   "repository": {
     "type": "git",
@@ -44,6 +44,7 @@
   "devDependencies": {
     "babel": "^5.8.23",
     "babel-eslint": "^4.1.3",
+    "babel-jest": "^5.3.0",
     "babelify": "^6.3.0",
     "browserify": "^11.2.0",
     "eslint": "^1.5.1",
@@ -61,6 +62,7 @@
     "watchify": "^3.4.0"
   },
   "jest": {
+    "scriptPreprocessor": "<rootDir>/node_modules/babel-jest",
     "unmockedModulePathPatterns": [
       "<rootDir>/node_modules/fbjs",
       "<rootDir>/node_modules/jest-cli",
